Fix #885: type coercion due to slightly incorrect default value (#1112)
When adding a new term to a `SymbolicOperator` with a sympy coefficient,
the coefficient was being cast to a float. This was because the default
value for a new term was `0.0`, which is a float. This commit changes
the default value to `0`, which is an integer. This ensures that the
type of the coefficient is preserved when adding new terms.